refactor(chatd): decimalize chat cost amounts end to end
Replace float-backed pricing (*float64) and int-backed total_cost_micros
(int64/sql.NullInt64) with github.com/shopspring/decimal across the Go
backend for exact fractional precision.

+++ coderd/database/dump.sql
@@ -1219,7 +1219,7 @@ CREATE TABLE chat_messages (
     context_limit bigint,
     compressed boolean DEFAULT false NOT NULL,
     created_by uuid,
-    total_cost_micros bigint
+    total_cost_micros numeric
 );
 
 CREATE SEQUENCE chat_messages_id_seq
